functional tests: enable 'CrossPackageNameClash' suite for Kotlin
This change enables 'CrossPackageNameClash' functional test
suite for Kotlin generator tests.

This test verifies if the types with conflicting names in LIME
files compile when Kotlin generator is used to process the input.

Note: this test fails. Kotlin generator does not generate valid code.
This will be fixed in next commit.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>