Kotlin: handle types with duplicated names
This change adds handling of types with duplicated names.
It ensurers that 'cross-package-name-clash' test suite
passes without problems.

Without this change Kotlin generator was creating multiple
imports of type with the same name. With this commit it does
not generate imports if there are types with duplicated names
and references them with full path.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>