Kotlin: reimplement static properties in interfaces
In order to use JvmField in companion object of
interface all properties/constants defined in
such companion object must be 'public final val'.

This restriction applies only to interface -- it
does not apply to classes.

Therefore, the implementation of static properties
has been changed -- instead of declaring val property
it is replaced with getter/setter functions.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>