KotlinConstant.mustache: use JvmField annotation for constants
This annotation is needed to access the constants from companion
objects in natural way when utilizing the Kotlin code from Java.
Without the annotation the get/set methods would need to
be used on companion.

With the annotation the constant can be used directly on class, not
companion (without get/set).

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>