Conditionally warn about android attributes mismatch (#1716)
In order to ease transition from Java to Kotlin
the users have the possibility to enable warnings
about mismatch in attributes for elements.

The following new CLI parameter is available
`-enableandroidattributesmismatchwarning` as well as
`GLUECODIUM_ENABLE_ANDROID_ATTRIBUTES_MISMATCH_WARNING`
CMake parameter.

When flag is set to true and an element from LIME file uses
annotations for Java, but it does not use the matching annotation
for Kotlin then the warning is generated (and vice versa).

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>