Add JVM annotations to avoid interoperability issues with Java (#1729)
------------------- Motivation -------------------
Java and Kotlin can be called from each other.
To avoid the interoperability problems withs Java and to
consume the Kotlin code from Java application(JVM),
certain annotations need to be used. Idea here is to add the
required annotations for operation.

------------------- Implemented solution -------------------
Following annotations are added to ensure proper operation.

1. `Throws` - To indicate a constructor/structure/functions throws a java checked exception.
2. `file:JvmName` - To give appropriate name for Kotlin top level elements in Java
3. `JvmOverloads` - To indicate functions with default values

Signed-off-by: amogbetkar <96105145+amogbetkar@users.noreply.github.com>