functional tests: adhere to LIME preconditions in Java tests
There were two tests, which explicitly passed 'null'
as a parameter to functions, which are defined as
taking non-optional parameters in LIME files. This
was invalid.

When such tests were run with Kotlin generated code
then JVM raised exception -- null passed for explicit
non-optional param.

This commit adjusts the mentioned tests and removes
the invalid logic.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>