functional tests: define 'JavaTestsRunWithKotlin' tag via CMake
The tag is needed to properly handle Java tests, which want to
validate types, which are only available for Java generator
e.g. via Skip/EnableIf or use Java name directly like JavaExternalTypes.

The new tag is used with 'EnableIf' for Kotlin to ensure, that
Kotlin generator provides required types to avoid compilation and runtime
errors.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>