KotlinNativeBase: add propagate exceptions flag (#1730)
Such flag is present in Java's version of NativeBase
and is used to control whether exceptions related to
removal of JNI wrappers are propagated to the caller.

This commit adds the flag also to KotlinNativeBase
in order to be able to use the generated Kotlin code
to run Java functional tests.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>