Kotlin: introduce usage of JvmField annotation (#1721)
This annotation is needed to access the constants from companion
objects and fields of structs in natural way when utilizing the generated Kotlin
code from Java. Without the annotation the get/set methods would need to
be used on companion and fields.

Note: this PR adjusts generation of static properties for interfaces. Sadly, JvmField
used in companion object of interface requires all properties to be public final val.
This restriction applies only to interface -- it does not apply to classes.
Therefore, the implementation of static properties has been changed -- instead of
declaring val property it is replaced with getter/setter functions.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>