Kotlin generator: temporarily revert usage of internal keyword
The internal keyword is quite problematic for interoperability
with Java and JNI. The usage of 'internal' forces the compiler
to perform additional name mangling. Such name mangling is not
trivial -- therefore, JNI function signatures do not match.

Moreover, when accessed from Java the generated functions and types
become public with additional name suffix. This is problematic.

This change disables generation of 'internal' keyword until
correct solution is found. The usage of 'JvmName()' was tested,
but it is not sufficient to cover all cases e.g. interfaces.

So, for the initial release of Kotlin generator the internal
keyword support is disabled.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>