functional tests: generate documentation for Kotlin code (#1717)
This change adjusts build scripts to use 'Dokka'
documentation engine. When the build script for
Kotlin functional tests is invoked with '--docs'
parameter, then a new gradle task called 'dokkaHtml'
is executed.

The mentioned task generates the documentation on
KDoc comments basis for all generated Kotlin files.
The output directory in which the documentation
will be available is 'build-android-kotlin/documentation'.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>