Kotlin: implement namerules tests (#1728)
In scope of name rules tests for Kotlin this change adds:
- the new shell script, that is used for building namerules tests for Kotlin generator
- the new name rules file for Kotlin tests: kotlin.properties
- the new test for Kotlin generator called NameRulesTest.kt

Moreover, this change adjusts gradle and CMake files to properly build the namerules tests for Kotlin.

Additionally, gluecodium-gradle plugin is extended to enable the users to run Kotlin generator instead
of Java generator. This is achieved via the new configuration properties.

Lastly, the CI workflow file for functional tests is extended with new step to run name rules tests for Kotlin.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>