functional tests: fix warnings for Kotlin build (#1741)
* ColorConverter.kt: remove compilation warnings

This is the custom input for functional tests
for Kotlin. The following warning was visible.

```
Unnecessary non-null assertion (!!) on a non-null
receiver of type Int
```

This seems to be false-positive -- removal of
non-null assertion caused compilation failure.
Moving the assertion above the function call
does not generate the warning. Therefore, such
way of fixing the warning is used.

* DatesTest.kt: do not use deprecated constructor

The following warning was visible:
```
'constructor Date(Int, Int, Int, Int, Int, Int)' is deprecated.
Deprecated in Java
```

This commit adjusts the construction of 'Date' objects
in the test to remove the warnings.

---------

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>