Kotlin: handle custom attributes (#1744)
---------- Motivation ----------
Kotlin generator is able to render valid Kotlin files. The handling includes
also all standard attributes/annotations, that are available for Java generator.
This is done to ensure the same level of support for both Kotlin and Java.

However, during implementation one small detail was missed. It is available in Java as:
```
@Java(Attribute=<SOME_CUSTOM_ATTRIBUTE>)
```

---------- Implemented solution ----------
Introduced KotlinAttributes and KotlinAttributesInline mustache templates and their
usage. The implemented feature is showcased in new case for smoke tests called 'attributes'.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>