CMake: handle usage of 'android' and 'android-kotlin' generators in one pass (#1750)
Initially, the usage of 'android' and 'android-kotlin'
generators in one pass was disallowed, because the generated
JNI symbols would cause compilation errors.

In order to fix the problem a new CMake parameter was added.
It is called 'GLUECODIUM_FORCE_USAGE_OF_JNI_FROM_JAVA_GENERATOR'.

When the optional property is set to true, then usage of both
generators is allowed. However, the generated Kotlin code is
going to use JNI layer generated by Java generator.

Therefore, it is crucial to ensure that LIME model for both
Java and Kotlin is same -- that's the responsibility of the user.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>