Kotlin: add 'JvmField' annotation to generated enumeration (#1754)
The annotation was missing in the case of 'value' field
of the generated enumeration. Moreover, the field was
private -- in Java generated code it is public.

This commit aligns the behavior of Kotlin generator
with Java generator. Furthermore, it adds a new test case
to functional tests -- it validates that the functionality works.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>