Add werror option for android attributes mismatch (#1755)
The mismatch between attributes for Kotlin and Java can
be easily found via '-enableandroidattributesmismatchwarning'.
The flag was added in previous release.

However, it would be nice to prevent developers from introducing
new places, which cause warnings. Therefore, the possibility to
set 'werror' flag for this check is added.

The unit and cmake tests are adjusted to confirm, that the functionality
works as expected.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>