smoke tests: add 'method overloads' test suite for Kotlin (#1723)
This change introduces the expected output
for smoke tests, which is aligned with the
same test suite for Java generator to ensure
that the level of support in Kotlin is the same.

This change adjusts the LIME file used as input for the
functional tests to match the one used in smoke tests.
This is done to ensure that the generated code compiles
and runs without problems.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>