Kotlin: handle accessors with custom platform names (#1759)
Accessors of a property may have custom names in LIME files.
This functionality works as expected for Java (android generator).

However, the tests revealed that it does not work correctly for Kotlin
generated code. In Kotlin JVM symbol for getter and setter is generated
by the compiler on property basis. Therefore, setting custom names in LIME
files causes mismatch between the name in our generated JNI layer and the
name expected by the Kotlin compiler.

This can be fixed via 'JvmName()' annotation for getters and setters in
Kotlin generated code. This commit brings handling of the described situation
in 'KotlinProperty.mustache'. When we identify the case in which getter/setter
needs custom name then 'JvmName()' is rendered. This allows the new functional
test to pass also for Kotlin.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>