Dart: generate type aliases
This change introduces generation of type aliases for
Dart. Updated other Dart templates and resolvers to
treat type aliases as a normal type, instead of
skipping it through to the target type.

This commit is just a rebase of another pull request.
Original author: Daniel Kamkha <daniel.kamkha@here.com>

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>