Validation: warn when unnamed parameters are documented for lambdas
When we do not specify lambda parameter names then 'p0, p1 ... pN'
are used. Such names do not bring any context for the caller to
understand their purpose.

Moreover, when such parameters are explicitly documented it is usually
a mistake. Therefore, 'LimeLambdaValidator' class is extended to print
warning or error (depending on 'werror' flag) when parameters without
custom names are explicitly documented for lambdas.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>