CI: run CMake tests for iOS (#1770)
---------- Motivation ----------
Gluecodium provides a set of CMake scripts, which allows users to integrate
code generation into their build system. Because of the number of parameters
and available configurations such scripts are tested via CMake unit tests.

There are some test cases, which are platform specific. For instance they are
just android-related or ios-related.

Because of the duration of jobs the pipeline executed on Gluecodium CI runs
CMake toolchain tests only for Android. Therefore, the test cases, which are
iOS-specific may start to fail – there is a chance that we will not catch the problem on CI.

All configured CI jobs finish in less than 10 minutes. CMake tests for iOS take 25-30 minutes
when run with multiple threads.

---------- Solution ----------

The new pipeline called cmake-tests-macos.yml is added to CI configuration files.
However, it is not executed for each and every change. It is triggered automatically only
when files from cmake directory are modified.

Moreover, the job may be triggered manually via GitHub actions tab, because the pipeline
is configured with workflow_dispatch event.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>