LimeValidatorUtils: handle non-trivial cases in 'needsDocumentationComment()' (#1771)
---------- Motivation ----------
The validation of missing documentation comments was introduced
a few months ago. It allows to raise a warning or error when the developer
forgets to document the public API.

The documentation is not needed for the internal API. Therefore, the internal
elements or elements nested in internal elements do not generate warnings.
This is achieved via 'needsDocumentationCommen()' function, which examines
global 'internal' attribute on elements.

It seems, that there exist cases in which this logic is not enough. For instance
let’s take a look at the following LIME input:

```
@Skip(Java, Kotlin, Dart) @Swift(Internal)
class SomeClass {
    // Factory to create instance of SomeClass on platform side.
    lambda SomeClassFactoryFromID = (ULong) -> SomeClass
}
```

The lambda is nested in internal class – it is skipped for all the platforms but Swift
and in Swift it is internal. We do not care about C++ because it is not directly used
by the clients. Therefore, we should not generate warnings for it.

---------- Solution ----------
Extended the function, which checks if the documentation is needed for the given
element to handle non-trivial cases and implemented the required unit tests.

Documentation comments are required for the given element when one of the following occurs:
1. The element is not internal and all of its parents are not internal.
2. The element is not internal for any non-skipped platform (JAVA, SWIFT, DART, KOTLIN) and its parents too.
3. The element is not skipped for all platforms and its parents too.
   If the element is skipped for all platforms it means that we generate only C++ -- for such case we do
   not need the documentation comment.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>