Dart/FFI: make callbacks queue aware of isolate context
The callback queue is created when Gluecodium's library context
is initialized in the Dart code. The callback queue processes
requests related to a particular isolate.

This change introduces a new code, which stores the information
about isolate context (its handle) in the newly created callback
queue object.

This handle will be used in cases when callbacks need to be executed
outside of the Flutter platform task. For instance from platform-specific
event handlers, which cannot return until some work is finished. In such
cases blocking the thread would stop processing of any callbacks, which
may cause dead-locks.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>