Dart/FFI: map isolate thread ID to its numeric ID
In callbacks queue manager class we keep the association
of callback queues to some numeric IDs.

However, each callback queue is associated with just one
isolate. In order to be able to process the callbacks
queue outside of Flutter task, but being in the same
thread we need to be sure, that the current thread is
associated with the given isolate.

This change introduces a new member variable to the manager
class. It contains the association of a particular thread
to the numeric value of ID for isolate.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>