Dart/FFI: allow processing callbacks queue from separate platform task (#1777)
---------- Motivation ----------
Before Flutter 3.29 the following threads were used by the Flutter engine:
- 'UI Thread' -- it executed Dart code
- 'Main Platform Thread' -- it executed platform specific callbacks

In Flutter 3.29 these threads were merged into one. The Dart code is executed
now by the 'Main Platform Thread'. The platform specific handlers are also executed
by that thread.

It is important to note that certain platform handlers cannot finish until other actions
are complete. For instance, the 'SurfaceHolder.Callback.surfaceDestroyed()' handler
on Android cannot finish until the separate rendering thread finishes using the resources.

From Android docs:
  "If you have a rendering thread that directly accesses the surface, you must ensure
  that thread is no longer touching the Surface before returning from this function. "

When such handler is executed in the 'Main Platform Thread' the processing of Dart
callbacks queue in Gluecodium does not happen. The whole thread is blocked. This
may lead to dead-lock if the thread, for which the handler is waiting tries to use the
Dart/Flutter callback from C++.

---------- Designed solution ----------
Implemented and exposed a new public function, which allows processing callbacks
queue for a given thread from a separate task (e.g. from platform specific handler).

With the usage of the new function the handlers may implement pooling approach.
In meantime they can process the callbacks queue. This way the risk of deadlock
is mitigated.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>