Dart/FFI: enter isolate only if callbacks queue is open (#1774)
There are corner cases in which the isolate, for which we
created a callback in the past may be closed. This can happen
when we register platform specific event handler (for instance
'appWillTerminate()' on iOS from Swift). With Flutter 3.29+
such callback is executed after the isolate is finished in some
environments.

In such case we should not enter the isolate manually in the
callback proxy. If we do it, then segmentation fault is going
to happen in internals of Dart DL API.

This change adjusts the code to check if the callbacks queue
for given isolate is open before entering it. In case when
it is closed, then the call is ignored.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>