Kotlin: reintroduce usage of internal keyword (#1760)
---------- Motivation ----------
The usage of 'internal' keyword in Kotlin generated code was disabled
in the past to progress with enabling Java tests and Java interoperability.

The 'internal' Kotlin keyword is quite problematic for interoperability with Java
and JNI layer generated by Gluecodium. The usage of 'internal' forces the compiler to
perform additional name mangling. Moreover, the 'internal' function in Kotlin becomes
public when consumed from Java.

Such name mangling is not trivial and causes the following problems:
- when internal is used for a function or type then JNI function signatures
   do not match -- when we try to call external internal function from Kotlin then
   it cannot be found in our shared library with C++ code
- when accessed from Java the generated functions and types become public
   with additional name suffix -- this is problematic and unexpected

---------- Implemented solution ----------
This change tries to ensure that 'internal' keyword can be used in Kotlin
and it works with the generated JNI layer. In order to ensure, that symbols are
not mangled the '@JvmName()' annotation is used. This way the generated code
can use internal keyword and still work properly.

Important: this change does not resolve the problem of public access to the
internal functions in Java. The possible fixes are being investigated.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>