functional: showcase compilation error
When the generated Java code is accessed from Kotlin
then single method interfaces are by default functional.

Because of that when anybody tries to switch from Java
to Kotlin generated code, he/she may encounter the following
error: single method interfaces in Kotlin are not annotated
as functional.

This functional test shows the compilation failure when
such usage is encountered.

In next commit the fix will be implemented.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>