Kotlin: make interfaces functional by default (#1780)
---------- Motivation ----------
When the generated Java code is accessed from Kotlin
then single method interfaces are by default functional.

Because of that when anybody tries to switch from Java
to Kotlin generated code, he/she may encounter the following
error: single method interfaces in Kotlin are not annotated
as functional by default.

---------- Solution ----------
Implemented the new predicate, which checks if the interface
can be annotated as functional and introduced the usage of
the new predicate in mustache template, which is used to generate
the Kotlin interfaces.

In Kotlin, a functional interface:
 - can have only one non-static function
 - cannot have non-static properties
 - can have multiple static functions/properties

This commit also introduces functional test, which was used to
showcase the compilation failure.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>