kotlin: adjust name of class for free functions
According to Kotlin documentation all the functions and properties
declared in a file (including extension functions) are compiled into
static methods of a Java class named after the file.

The name can be adjusted using 'file:JvmName()' annotation.
Because initially the value for JvmName of file-level declarations
was adjusted to be the same as the type inside, when we declared
extension function then we would have two classes with the same name.

In fact, because we did not have any free functions/properties in the
generated code the value of 'file:JvmName()' was not used.

This commit adjusts the name to have 'Extensions' suffix to avoid
compilation problems. The extensions are needed to ensure that the
provided test case compiles.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>