CI: fix 'Swift for MacOS' job
The job was failing with the following message:
 - 'Adding XCTest bundles requires CMAKE_OSX_SYSROOT to be set.'

It was caused by the upgrade of CMake on the runner image to version 4.+.

The fix was implemented using CMake documentation.
The variable is initialized to value of 'SDKROOT' environment
variable. Moreover, the deployment target is set manually via
'MACOSX_DEPLOYMENT_TARGET' to avoid execution failures.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>