Kotlin: explicit accessors for static properties (#1782)
---------- Motivation ----------
Java generator in Gluecodium generates getter/setter methods for static properties.
When such Java code is accessed from Kotlin the methods should be represented as
Kotlin synthetic properties (according to the documentation of Kotlin language).

However, it turns out that such static properties can be also accessed via getter and setter
static functions. Such usage was identified to be working.

Therefore, in order to allow seamless replacement of Java generated code with Kotlin generated
code for Kotlin apps the Kotlin generated code must also allow getting/setting static properties
via getter/setter functions.

---------- Implemented solution ----------
Kotlin generator produces additional extension methods for companion object to allow getting and
setting static properties. The generated methods are annotated as JvmSynthetic to ensure that
they are not visible in Java (because such feature is Kotlin-specific).

Moreover, the value passed to file:JvmName() is adjusted to avoid compilation problems. That name
is used only when we have global functions or properties. Such entities are compiled to the class with
specified name. Because we used the same name as the element defined for given file, when extensions
are defined we would end up with two classes with the same name -- this was adjusted.

Signed-off-by: Patryk Wrobel <183546751+pwrobeldev@users.noreply.github.com>